chatbot/requirements.txt
Flask
dapr
python-telegram-bot==13.7
requests
psycopg2-binary
